Move test utils to general test folder out of test_views
