+++ tests/test_views/test_analysis.py
@@ -1,7 +1,7 @@
 from django.test import Client, TestCase
 from django.urls import reverse
 
-from tests.test_views.utils import TestConfiguration
+from tests.utils import TestConfiguration
 
 
 class AnalysisTestCase(TestCase):
