+++ tests/test_views/test_core_background_tasks.py
@@ -1,7 +1,7 @@
 from django.test import TestCase, Client
 
 from core.models import CombineBackgroundTask
-from tests.test_views.utils import TestConfiguration
+from tests.utils import TestConfiguration
 
 
 class BackgroundTaskTestCase(TestCase):
