+++ tests/test_views/test_external_background_tasks.py
@@ -1,7 +1,7 @@
 from django.test import Client, TestCase
 
 from core.models import LivySession
-from tests.test_views.utils import TestConfiguration
+from tests.utils import TestConfiguration
 
 
 class ExternalBackgroundTaskTestCase(TestCase):
