+++ tests/test_views/test_oai_endpoint.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 from django.core.exceptions import ObjectDoesNotExist
 
 from core.models import OAIEndpoint
-from tests.test_views.utils import TestConfiguration
+from tests.utils import TestConfiguration
 
 
 class OAIEndpointTestCase(TestCase):
