+++ tests/test_views/test_record.py
@@ -1,7 +1,7 @@
 from django.test import Client, TestCase
 
 from core.models import ValidationScenario
-from tests.test_views.utils import TestConfiguration
+from tests.utils import TestConfiguration
 
 
 class RecordTestCase(TestCase):
