+++ tests/test_views/test_stateio.py
@@ -1,7 +1,7 @@
 from django.test import Client, TestCase
 
 from core.models import StateIO
-from tests.test_views.utils import TestConfiguration
+from tests.utils import TestConfiguration
 
 
 class StateIOTestCase(TestCase):
